compiler: move Optimizer function to transform package
This refactor is a prerequisite to much larger refactors in the
compiler.
